7.5 update

+++ WrathCombo/WrathCombo.cs
@@ -302,7 +302,7 @@ public sealed partial class WrathCombo : IDalamudPlugin
         }
     }
 
-    private void ClientState_TerritoryChanged(ushort obj)
+    private void ClientState_TerritoryChanged(uint obj)
     {
         UpdateCaches(false, true, false);
